requirements.txt
customtkinter ~= 5.2.0
pillow ~= 9.5.0
